Fix client tests and improve coverage to 89% (from 34%)
- Fixed test_client.py to use conftest fixtures with JSON codec setup
- Fixed retry_job() SQL query - added ::bigint casts for parameter type inference
- Fixed migration 005 and 008 with explanatory comments about notices
- Updated AGENT_DEVELOPER_STARTUP.md with critical su - command for DB setup
- Fixed setup-test-db-nosudo.sh to handle unbound USER variable

Test Results:
- All 52 client tests now passing (was 23 failing)
- client.py coverage: 34% → 89% (improved 55%)
- test_client.py: 36 tests passing
- test_client_core.py: 16 tests passing

Key Fixes:
1. asyncpg 0.30.0 JSON codec - tests now use db_pool from conftest
2. PostgreSQL type inference - added ::bigint casts in retry_job query
3. Migration notices documented as expected/safe idempotent behavior

+++ pyjobby/client.py
@@ -556,10 +556,10 @@ class JobClient:
                     job_class, kwargs, queue, prio, uid, capability,
                     TIMEZONE('utc', clock_timestamp()) as run_after,
                     run_group,
-                    (COALESCE(admin_data::text::jsonb, '{}'::jsonb) || jsonb_build_object('retry_of', $1))::json as admin_data,
+                    (COALESCE(admin_data::text::jsonb, '{}'::jsonb) || jsonb_build_object('retry_of', $1::bigint))::json as admin_data,
                     'queued' as state
                 FROM jorb
-                WHERE id = $1
+                WHERE id = $1::bigint
                   AND state IN ('crashed', 'finished')
                 RETURNING id
             """, job_id)
